heb de flow van de kamers ge updadet en en buggse er uit gehaalt, ook de ui voor debuging bij gewerkt

+++ FRDM_MCXA153/robox/state_logic/game_data.h
@@ -47,7 +47,7 @@ typedef struct
 typedef struct 
 {
     uint8_t difficulty;
-    uint8_t totalTime;
+    float totalTime;
     bool audio;
 
 }globalSettings_t;
